[Ionut] [[BR]] - Implemented a basic ThirdPerson camera [[BR]] -- Does not limit rotations yet [[BR]] -- Does not have on click and drag functionality yet [[BR]]
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@241 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 8208d7138400cb8b6d085ef4df422f0b2d2416d1 [formerly b97cc26441abe3f548a3b16eba33ea1fae39318c]
Former-commit-id: ddd299082e49175ae6bd9631bea075a2d6decac9

+++ Source Code/Hardware/Video/OpenGL/GLFWWrapper.cpp
@@ -561,8 +561,8 @@ void GL_API::setWindowPos(GLushort w, GLushort h) const {
     glfwSetWindowPos(Divide::GL::_mainWindow,w,h);
 }
 
-void GL_API::setMousePosition(GLdouble x, GLdouble y) const {
-    glfwSetCursorPos(Divide::GL::_mainWindow,x,y);
+void GL_API::setMousePosition(GLushort x, GLushort y) const {
+    glfwSetCursorPos(Divide::GL::_mainWindow,(GLdouble)x,(GLdouble)y);
 }
 
 I32 GL_API::getFont(const std::string& fontName){
